Make flat file disassemnly stay within bounds.
This makes the assembly of flat files stay within the file loaded and
not try to trace the rest of the address space.  To do that it
implements a base address so the code can be loaded into the internal
array starting a location zero but the addresses presented are relative
to that base address.  The 6502 disassembler also has to take account of
this base address when handling absolute addresses.  CINTCODE uses
completely relative addressing.