Command line args, ROM symbols and bug fixes.
This enables loading of the BCPL ROM symbols from a built-in table and
prepares the way for user-loadable symbols (from a label file).

In doing so it introduces formal command line argument parsing.  It also
re-arranges the way files are loaded so there are fixed 64K arrays for
the file content and the index and files are loaded at their natural
base address rather than at zero.

Finally this fixes some bugs in tracing both CINTCODE and 6502.